001 tasks extracted and properly executed. Properties were previously written during configuration
